refactor(utils): consolidate shallow single-purpose modules
- Move TEX_TOOLS constant into platformPaths.ts (only consumer)
  - Delete utils/system/constants.ts (was 2 lines)

- Move SETTINGS_QUERY into config/constants.ts
  - Delete utils/settingsQueries.ts (was 7 lines)
  - Update imports in settingsCommands.ts, MainViewMessageHandler.ts,
    SettingsManager.ts to use @utils/config

This reduces file count by eliminating modules that existed
only to export a single constant or object.
