+++ src/utils/config/constants.ts
@@ -1,6 +1,15 @@
 // Local imports - config utils
 import { getConfig } from './configUtils';
 
+// Settings query constants for VS Code settings UI
+export const SETTINGS_QUERY = {
+  EXTENSION: '@ext:texra-ai.texra',
+  WORKFLOW_AGENTS: '@ext:texra-ai.texra texra.agents',
+  TOOL_USE_AGENTS: '@ext:texra-ai.texra texra.toolUseAgents',
+  MODELS: '@ext:texra-ai.texra models',
+  AGENT_DIRECTORY: '@ext:texra-ai.texra explorer.agentsDirectory',
+} as const;
+
 // Common file type constants
 export const FILE_TYPES = [
   'input',
