--- src/utils/system/constants.ts
@@ -1,2 +1,0 @@
-// Common LaTeX tool names used across the system
-export const TEX_TOOLS = ['latexdiff', 'latexindent', 'latexmk'] as const;
