+++ src/utils/system/platformPaths.ts
@@ -9,8 +9,8 @@ import { execaSync } from 'execa';
 import * as logger from '@logger/logUtils';
 import { AbsoluteFS } from '@utils/files';
 
-// Local file imports
-import { TEX_TOOLS } from './constants';
+// Common LaTeX tool names used across the system
+const TEX_TOOLS = ['latexdiff', 'latexindent', 'latexmk'] as const;
 
 const CHANNEL = 'platformPaths';
 logger.initialize(CHANNEL);
