+++ src/webview/MainViewMessageHandler.ts
@@ -12,13 +12,12 @@ import { MAIN_VIEW_COMMANDS } from '@common/webview';
 import { agentDirectories } from '@frontend/agents';
 import { showInstructionWithSuppress } from '@frontend/ui/instruction';
 import { computeModelOptions } from '@model/computeModelOptions';
-import { getConfig, setConfig } from '@utils/config';
+import { getConfig, setConfig, SETTINGS_QUERY } from '@utils/config';
 import {
   safeExecuteCommand,
   checkCoreDependencies,
   getToolDocsCommand,
 } from '@utils/system';
-import { SETTINGS_QUERY } from '@utils/settingsQueries';
 import { AUTH_COMMANDS, getAuthStatus } from '@commands/auth';
 import { PROVIDER_URLS } from '@commands/api/apiKeyCommands';
